==> ML/1.TMDBOpenAPI.ipynb <==
<pre>TMDB 오픈 API를 가져와서 데이터가 없는 값(결측치)은 제거</pre>

<pre>오픈 API 에서 가져온 데이터 영화제목, 포스터 URL, 개봉시기, 예산, 장르, 국가, 연출차, 감독, 상영시간, 소개, 평점, 투표횟수</pre>

# TMDB Open API : TMDB API를 가져오는 파일
import requests
import json

# json 에서 API Key값 가져오기 : key, token
file_path = "./APIKey/TMDB.json"
with open(file_path, 'r') as file:
    api = json.load(file)

key = api['key']
token = api['token']

movies = "616037"
# movie info...
url = f"https://api.themoviedb.org/3/movie/{movies}?api_key={key}"
req = requests.get(url)
api_info = json.loads(req.text)

# title str값 => id값으로 정수 부여
# genres dict -> key(int 'id') : value(str 'name')
# country : 다른 것도 같이 확인 필요!!!
title = api_info['title']
country = api_info['production_countries'][0]['iso_3166_1']
overview = api_info['overview']

print("title: ", title, ", country: ", country)
print("overview: ", overview)
print("genres: ", api_info['genres'])

# cast, crew ...
url = f"https://api.themoviedb.org/3/movie/{movies}/credits?api_key={key}"
req = requests.get(url)
api_people = json.loads(req.text)

for person in api_people['crew']:
    try :  
        if(person['job'] == 'Director'):
            print(person)
    except:
        pass 

# id And title
# total_pages : 34599
# total_data : 691978
page = 1
url = f"https://api.themoviedb.org/3/discover/movie?page={page}&api_key={key}"
req = requests.get(url)
api_movie = json.loads(req.text)

for movie in api_movie['results']:
    print(movie['id'], movie['title'])

- 영화 : 제목, 평가, 투표수, 장르, 국가, 개봉일, 감독, 출연, 등급, 상영시간, 줄거리, 이미지 경로
- 머신러닝 : 제목, 예산, 평가, 투표수, 장르, 국가, 감독, 출연

import numpy as np
from tqdm import tqdm
pages = np.arange(500) + 1

movie_list = []
for page in tqdm(pages):
    url = f"https://api.themoviedb.org/3/discover/movie?page={page}&api_key={key}&language=ko-KR"
    req = requests.get(url)
    api_movie = json.loads(req.text)
    for movie in api_movie['results']:
        movie_id = []
        movie_id.append(movie['id'])
        for movie in movie_id:
            url = f"https://api.themoviedb.org/3/movie/{movie}?api_key={key}&language=ko-KR"
            req = requests.get(url)
            api_info = json.loads(req.text)
            title= {movie : api_info['title']}
            try:
                poster_path = "https://image.tmdb.org/t/p/original" + api_info['poster_path']
            except:
                poster_path = ""
            release_date = api_info['release_date']
            budget = api_info['budget']
            genres = api_info["genres"]
            country = api_info['production_countries']
            vote_average = api_info['vote_average']
            vote_count = api_info['vote_count']
            runtime = api_info['runtime']
            overview = api_info['overview']
            url = f"https://api.themoviedb.org/3/movie/{movie}/credits?api_key={key}&language=ko-KR"
            req = requests.get(url)
            api_people = json.loads(req.text)
            for actor in api_people['cast']:
                try:
                    if(actor['order'] == 1):
                        main_charactor = {actor['id'] : actor['name']}
                except:
                    pass
            for crew in api_people['crew']:
                try :  
                    if(crew['job'] == 'Director'):
                        director = {crew['id'] : crew['name']}
                except:
                    pass 
            movie_list.append([title, poster_path, release_date, budget, genres, country,
                              main_charactor, director, runtime, overview, vote_average, vote_count])
            #del title, budget, genres, country, vote_average
            director = "-"
        del movie_id


import pandas as pd
df = pd.DataFrame(movie_list, columns="title, poster_path, release_date, budget, genres, country, main_charactor, director, runtime, overview, vote_average, vote_count".split(", "))
df.to_csv("data/movie_data_original.csv", index=False)

movie_df = pd.read_csv(
    "data/movie_data_original.csv", names="title, poster_path, release_date, budget, genres, country, main_charactor, director, runtime, overview, vote_average, vote_count".split(", "), skiprows = 1)
print("total movie_data_original: ", len(movie_df))
movie_df.head(1)

# 감독이 없는 데이터 87개 제거
NoDirector = movie_df[movie_df['director'] == '-']
movie_df1 = movie_df.drop(NoDirector.index)
len(movie_df1)

# 예산이 없는 데이터 4,604개 제거
NoBudget = movie_df1[movie_df1['budget'] == 0]
movie_df2 = movie_df1.drop(NoBudget.index)
len(movie_df2)

# 장르가 없는 데이터 2개 제거
movie_df2 = movie_df2.drop(movie_df2[movie_df2['genres'].apply(lambda x: len(x) == 2)].index)
len(movie_df2)

# 국가별이 없는 데이터 5개 제거
movie_df2 = movie_df2.drop(movie_df2[movie_df2['country'].apply(lambda x: len(x) == 2)].index)
len(movie_df2)

movie_df2.to_csv("data/movie_data_final.csv", index=False)

# 머신러닝 : 제목, 예산, 평가, 투표수, 장르, 국가, 감독, 출연
movie_df2 = movie_df2[['title', 'budget', 'genres','country', 'main_charactor', 'director', 'vote_average', 'vote_count']]

movie_df2.to_csv("data/movie_data_ML.csv", index=False)

# 결측치가 있는 데이터들... 네이버 영화 검색 api로 채우기?
