==> ML/2.TMDBCrawling.ipynb <==
<pre>TMDB 결측치값중 크롤링 할 수 있는 데이터를 가져와서 데이터를 채운 파일</pre>

# 데이터가 부족한 경우 사용
# NaN값을 제거하면 데이터가 5302개이므로 클롤링을 통해 NaN값을 채울려고 한다.
import pandas as pd
import numpy as np

- 데이터 분석(결측치 분석) 감독, 예산, 국가, 장르

# title budget genres country main_charactor director vote_average vote_count
df = pd.read_csv('data/movie_data_original.csv',
                 names="title postpath openDate budget genres country main_charactor director time overview vote_average vote_count".split(),
                 skiprows=1)

# 예산이 표기된 api가 없다...일단 제거 (4858개)
NoBudget = df[df['budget'] == 0]
df = df.drop(NoBudget.index)
len(df)

# 감독이 없는 데이터 4개
NoDirectors = df[df['director'] == '-']
len(NoDirectors)

# 감독이 없는 리스트 중 타이틀을 string -> dict로 그중 title name만 가져옴
# Ndirector 에 리스트로 저장
import ast
Ndirector = []
for NoDirector in df.loc[NoDirectors.index].title:
    for s in ast.literal_eval(NoDirector).values():
        Ndirector.append(s)
Ndirector

# Mr. Sadman 검색 결과 나오지 않음
# The Collective: Movie => 알레그잔데르 나나우 감독
# 아래 데이터는 IMDB에서 결과가 나옴.
# The Visual Effects of 'Scary Movie 4' => Matthew Asner
# Dragon Ball Z: The Fall of Men => 요한 포레

# 장르 4개
Nogenre = df[df['genres'] == '[]']
len(Nogenre)

# 감독이 없는 리스트 중 타이틀을 string -> dict로 그중 title name만 가져옴
# Ndirector 에 리스트로 저장
Ngenre = []
for Nog in df.loc[Nogenre.index].title:
    for g in ast.literal_eval(Nog).values():
        Ngenre.append(g)
Ngenre

# Mr. Sadman 검색 결과 나오지 않음
# 아래 데이터는 IMDB에서 결과가 나옴.
# The Visual Effects of 'Scary Movie 4' => Documentary
# 몽키킹: 전설의 시작 => 액션 / 판타지
# Die Taskforce-Nachbarn => Comedy

# 국가 4개
Nocountry = df[df['country'] == '[]']
len(Nocountry)

Ncountry = []
for Noc in df.loc[Nocountry.index].title:
    for c in ast.literal_eval(Noc).values():
        Ncountry.append(c)
Ncountry

# Mr.Sadman 결과는 없음 삭제
del Ncountry[4]

# 국가가 없는 영화중 개봉시기가 없는 영화가 있다.
Cdate = []
for Noc in df.loc[Nocountry.index].title:
    date = str(df.loc[df.title == f"{Noc}"]['openDate'])
    Cdate.append(date[8:12])
Cdate

# Mr.Sadman 결과는 없음 삭제
del Cdate[4]

import requests
import json
from tqdm.notebook import tqdm

url = f"https://imdb-api.com/en/API/Title/k_n5qxjb05/tt1375666"
response = requests.get(url)
result = json.loads(response.text)

<pre>IMDB 하루에 100번 밖에 호출이 불가능 조심할것</pre>

# IMDB 사용 제목으로 검색후 ID값 받기
movieid = []
movie_country = {}
for idx, title in tqdm(enumerate(Ncountry)):
    url = f'https://imdb-api.com/en/API/SearchMovie/k_n5qxjb05/{title}'
    response = requests.get(url)
    result = json.loads(response.text)
    try:
        for res in result['results']:
            if res['description'] == f"({Cdate[idx]})":
                movieid.append(res[0]['id'])
        print(title, movieid[idx])
        for code in movieid:
            url = f"https://imdb-api.com/en/API/Title/k_n5qxjb05/{code}"
            response = requests.get(url)
            result = json.loads(response.text)
            try:
                movie_country[title] = result['countries']
            except:
                print(title, " 의 국가가 나오지 않습니다.")
    except:
        print(title, " 의 id값이 없습니다.")

# Night Feeders => USA
# Shamelessly She-Hulk => USA
# Spider-Man: Convergence(2012) => USA
# Mirreyes vs Godínez => 멕시코
# Mr. Sadman => 결과값 x
# Perfection => USA
# Perfection => USA
# Red Hook => USA
# The Visual Effects of 'Scary Movie 4' => USA
# 몽키킹: 전설의 시작 => China
# 인비저블걸 => USA
# Dragon Ball Z: The Fall of Men => France
# Die Taskforce-Nachbarn => German

- 클롤링(네이버 사용X)

# 네이버 api에서 제목 검색 => 실패
with open('APIKey/Naver.json') as file:
    api = json.load(file)
client_id = api['id']
client_secret = api['secret']

search = "Chip and Potato: Chip's Holiday"
url = "https://openapi.naver.com/v1/search/movie.json"
headers = {
            'X-Naver-Client-Id': client_id,
            'X-Naver-Client-Secret': client_secret
           }
params = {
    'query' : '[극장판] 레이디버그: 미라클스톤의 비밀'
}
response = requests.get(url, headers=headers, params=params)
response

result = json.loads(response.text)
result
