==> ML/3.TMDBData.ipynb <==
<pre>투표 가중치(vote_average)를 새로 만들고 데이터 전처리(id값과 name값으로 나눔, 필요없는 데이터는 제거)를 한 파일</pre>

<pre>
# vote_average를 가중치를 부여
# vote = (v/(v+m))*R + (m/(v+m))*C
# v : 개별 영화에 투표한 횟수, m : 평점을 부여하기 위한 최소 투표 횟수
# R : 개별 영화에 대한 평균 평점(=vote_average), C : 전체 영화에 대한 평균 평점

# m : 투표 횟수의 평균 * 60%
# C : 평균 평점 * 60 %
</pre>

# TMDBML : TMDB 머신러닝 모델 만들기
import pandas as pd
import numpy as np

df = pd.read_csv("data/movie_data_final.csv")
df.head(1)  # title : str

# m : 투표 횟수의 평균 * 60%
# C : 평균 평점 * 60 %
m = df.vote_count.mean() * 60 / 100
C = df.vote_average.mean() * 60 / 100
print("m : ", m)
print("C : ", C)

v = df.vote_count.astype(int)
R = df.vote_average.astype(float)
df['result'] = (v/(v+m))*R + (m/(v+m))*C
for idx in df.index:
    if df.vote_count[idx] == 0 or df.vote_average[idx] == 0:
        df['result'][idx] = 0

import ast #형태가 리스트/딕셔너리인 문자열을 다시 리스트/딕셔너리 형태로 바꿀 수 있다.

# title: 이름, id/ genres: to_categorical/ country: to_categorical/ charactor: 이름, id/ director: 이름, id/ result 0 or 1
# numpy 배열 (5302,)
# df['genres'] = df['genres'].apply(literal_eval) 로 다시 짤 수 있다.
df_list = []
title, budget, genres, country, charactor, director, result = [], [], [], [], [], [], []
for idx in df.index:
    title.append([ast.literal_eval(df.title[idx])])
    budget.append([int(df.budget[idx])])
    genres.append(ast.literal_eval(df.genres[idx]))
    country.append(ast.literal_eval(df.country[idx]))
    charactor.append([ast.literal_eval(df.main_charactor[idx])])
    director.append([ast.literal_eval(df.director[idx])])
    result.append([1 if df.vote_average[idx] >= 0.5 else 0])
df_list.append([title, budget, genres, country, charactor, director, result])
df_arr = np.array(df_list)

# for title, charactor, director
class SeperateID_Name:
    def __init__(self, column):
        self.id = []
        self.name = []
        self.column = column

    def print(self):
        for col in self.column:
            for key, value in col[0].items():
                self.id.append(int(key))
                self.name.append(str(value))
        return self.id, self.name

# title : dictionary about title_ID & title_Name
sep_title = SeperateID_Name(title)
title_ID, title_Name = sep_title.print()
# charactor : dictionary about char_ID & char_Name
sep_charactor = SeperateID_Name(charactor)
char_ID, char_Name = sep_charactor.print()
# director : dictionary about dir_ID & dir_Name
sep_director = SeperateID_Name(director)
dir_ID, dir_Name = sep_director.print()

# country dictionry name to id 만들기
country_data = []
for cs in country:
    for c in cs:
        country_data.append(c['name'])
country_dict = {}
i = 0
for l in country_data:
    try:
        country_dict[l]  # 찾으면 패스 못찾으면 선언 및 아이디 부여
    except:
        country_dict[l] = i
        i += 1

# for genres, country
class SeperateID_Name_list:
    def __init__(self, column):
        self.ids, self.id = [], []
        self.names, self.name = [], []
        self.column = column

    def print(self):
        for cols in self.column:
            for col in cols:
                self.id.append(list(col.values())[0])
                self.name.append(list(col.values())[1])
            self.ids.append(self.id)
            self.names.append(self.name)
            self.id = []
            self.name = []
        return self.ids, self.names

# genres
sep_genres = SeperateID_Name_list(genres)
genre_ID, genre_Name = sep_genres.print()
# country
sep_country = SeperateID_Name_list(country)
country_ID, country_Name = sep_country.print()

# country_ID id 저장
country_I = []
country_ID = []
for names in country_Name:
    for name in names:
        country_I.append(country_dict[name])
    country_ID.append(country_I)
    country_I = []

- Id로 이루어진 데이터프레임(머신러닝용1)

df.head(1)

# title	budget	genres	country	main_charactor	director	vote_average	vote_count	result
df_ID = pd.DataFrame({'title': title_ID, 'budget':df.budget, 'genres':genre_ID, 'country':country_ID, 'charactor': char_ID, 'director':dir_ID, 'vote_score':df.vote_average, 'vote_count':df.vote_count,
'vote_weight':df.result})

df_ID.head(1)

df_ID.to_csv('data/movie_data_ID.csv', index=False)

- Name으로 이루어진 데이터프레임(분석용)

# title	budget	genres	country	main_charactor	director	vote_average	vote_count	result
df_Name = pd.DataFrame({'title': title_Name, 'budget':df.budget, 'genres':genre_Name, 'country':country_Name, 'charactor': char_Name, 'director':dir_Name, 'vote_score':df.vote_average, 'vote_count':df.vote_count,
'vote_weight':df.result})

df_Name.head(1)

df_Name.to_csv('data/movie_data_Name.csv', index=False)

- Id로 이루어진 데이터프레임(머신러닝용2)

- genres, country columns

genres_list = ["액션", "모험" ,"판타지", "가족", "애니메이션", "코미디", "SF", "드라마", "공포", "스릴러", "범죄", "미스터리", "음악", "역사", "TV 영화", "전쟁", "로맨스", "다큐멘터리", "서부"]

# genres 종류에 따라 나온 결과 값을 데이터프레임 변형, 해당하는 값만 1로 부여하고 나머지는 0으로 부여
# 가중치 값은 반올림하여 정수형으로 바꿈.
df_genre_each = pd.DataFrame(columns=genres_list)
for idx in df.index:
    genres = df_Name.genres[idx]
    # 한 줄 0으로 초기화
    df_genre_each.loc[idx] = 0
    for genre in genres:
        df_genre_each.loc[idx][str(genre)] = 1  # 한 줄 값 부여

df_genre_each.to_csv("data/movie_genres_data.csv", index=False)
