==> ML/6.TMDBMLuse.ipynb <==
<pre>머신러닝 모델 개발된 모델을 기존 영화들 흥행예측에 사용한 파일</pre>

import pandas as pd
from tensorflow.keras.models import load_model
import numpy as np

<pre>1. 데이터 불러오기</pre>

df = pd.read_csv('data/movie_data_ML_op.csv')

df_title = df[['title', 'vote_weight']]

df.drop(['title'], axis=1, inplace=True)

X = np.array(df)

y = X[:, 13:]
X = X[:, :-1]

print(X.shape)
print(y.shape)

<pre>2. 모델 불러오기</pre>

model = load_model('model/Dense_Dropout.h5')

yhat = model.predict(X)

yhat.shape

df_title['predict'] = yhat

df_title.to_csv('data/movie_data_predict.csv', index=False)

df_title.predict.mean()
