==> ML/7.Moviepredict.ipynb <==
<pre>개봉예정영화 네이버에서 크롤링한 파일</pre>

<pre>영화 개봉 예정 흥행 예측후 db로 저장</pre>

- 네이버 개봉예정영화(72개) 클롤링

from selenium import webdriver
from selenium.webdriver.common.by import By
import time
from bs4 import BeautifulSoup
from tqdm.notebook import tqdm

url = 'https://search.naver.com/search.naver?where=nexearch&sm=tab_etc&qvt=0&query=%EA%B0%9C%EB%B4%89%EC%98%88%EC%A0%95%EC%98%81%ED%99%94'
driver = webdriver.Chrome('APIKey/chromedriver')
driver.get(url)
time.sleep(2)

- table은 영화개봉예정 page당 8개의 영화로 구성되어있다.

driver.find_element(By.CSS_SELECTOR, 'div.cm_paging_area > .pgs > .pg_next').click()

table = driver.find_elements(By.CSS_SELECTOR, '.cm_content_wrap .card_item')

page_html = BeautifulSoup(driver.page_source, 'html.parser')
table_html = page_html.select('.cm_content_wrap .card_item')

table_url = []
table_url.append(table_html[0].select_one('.this_text')['href'])

url = 'https://search.naver.com/search.naver' + table_url[0]
url

table[0].find_element(By.CSS_SELECTOR, '.this_text').click()
time.sleep(2)

- 영화를 각각 클릭후 기본정보, 감독/출연, 포토를 클릭하여 가져온다.

- 액션	모헌	판타지	SF	드라마	공포	스릴러	미스터리	로맨스	budget	charactor	director	vote_weight

soup = BeautifulSoup(driver.page_source, 'html.parser')

# 전체 > 포스터
poster_path = soup.select_one('img._img')["src"]

# 기본 정보: 1, 감독/출연: 2, 포토: 5
# 기본 정보: 1
driver.find_elements(By.CSS_SELECTOR, '.tab_list > li')[1].find_element(By.CSS_SELECTOR, 'a').click()

driver.find_elements(By.CSS_SELECTOR, 'dl.info .info_group')[1].find_element(By.CSS_SELECTOR, 'dt').text

# 기본정보 > 장르
infos = driver.find_elements(By.CSS_SELECTOR, 'dl.info .info_group')
for info in infos:
    if(info.find_element(By.CSS_SELECTOR, 'dt').text == "장르"):
        genres_str = info.find_element(By.CSS_SELECTOR, 'dd').text
geres = genres_str.split(', ')

# 기본정보 > 소개
introduction = driver.find_element(By.CSS_SELECTOR, 'p._content_text').text

# 감독/출연: 2
driver.find_elements(By.CSS_SELECTOR, '.tab_list > li')[2].find_element(By.CSS_SELECTOR, 'a').click()

# 감독/출연 > 감독: 0
director = driver.find_elements(By.CSS_SELECTOR, 'div.cast_box')[0].find_element(By.CSS_SELECTOR,'ul.cast_list span._text').text

# 감독/출연 > 출연: 1
mainCharactor = driver.find_elements(By.CSS_SELECTOR, 'div.cast_box')[1].find_elements(By.CSS_SELECTOR,'ul.cast_list li')[0].find_element(By.CSS_SELECTOR, 'span._text').text

<pre>크롤링</pre>

movie = []
for page in tqdm(range(9)):
    url = 'https://search.naver.com/search.naver?where=nexearch&sm=tab_etc&qvt=0&query=%EA%B0%9C%EB%B4%89%EC%98%88%EC%A0%95%EC%98%81%ED%99%94'
    driver = webdriver.Chrome('APIKey/chromedriver')
    driver.get(url)
    time.sleep(2)
    for click in range(page):
        driver.find_element(By.CSS_SELECTOR, 'div.cm_paging_area > .pgs > .pg_next').click()
        time.sleep(1)
    page_html = BeautifulSoup(driver.page_source, 'html.parser')
    table_html = page_html.select('.cm_content_wrap .card_item')
    table_url = []
    for idx in range(8):
        try:
            table_url.append(table_html[idx].select_one('.this_text')['href'])
        except:
            pass
    driver.close()
    for idx in range(len(table_url)):
        url = 'https://search.naver.com/search.naver' + table_url[idx]
        driver = webdriver.Chrome('APIKey/chromedriver')
        driver.get(url)
        time.sleep(2)
        # 포스터 url Path
        soup = BeautifulSoup(driver.page_source, 'html.parser')
        try:
            poster_path = soup.select_one('img._img')["src"]
        except:
            poster_path = ""
        driver.find_elements(By.CSS_SELECTOR, '.tab_list > li')[1].find_element(By.CSS_SELECTOR, 'a').click()
        # 기본정보 > 장르
        infos = driver.find_elements(By.CSS_SELECTOR, 'dl.info .info_group')
        for info in infos:
            if(info.find_element(By.CSS_SELECTOR, 'dt').text == "장르"):
                genres_str = info.find_element(By.CSS_SELECTOR, 'dd').text
        geres = genres_str.split(', ')
        # 기본정보 > 소개
        try:
            introduction = driver.find_element(By.CSS_SELECTOR, 'p._content_text').text
        except:
            introduction = ""
        # 감독/출연: 2
        driver.find_elements(By.CSS_SELECTOR, '.tab_list > li')[2].find_element(By.CSS_SELECTOR, 'a').click()
        # 감독/출연 > 감독: 0
        try:
            director = driver.find_elements(By.CSS_SELECTOR, 'div.cast_box')[0].find_element(By.CSS_SELECTOR,'ul.cast_list span._text').text
        except:
            director = ""
        # 감독/출연 > 출연: 1
        try:
            mainCharactor = driver.find_elements(By.CSS_SELECTOR, 'div.cast_box')[1].find_elements(By.CSS_SELECTOR,'ul.cast_list li')[0].find_element(By.CSS_SELECTOR, 'span._text').text
        except:
            mainCharactor = ""
        movie.append([poster_path, geres, director, mainCharactor, introduction])
        driver.close()

import pandas as pd

df = pd.DataFrame(movie, columns=['poster_path', 'genres', 'director', 'mainCharactor', 'introduction'])

# null 값 확인
df.drop(df[df.poster_path == ""].index, inplace=True)
df.drop(df[df.genres == ""].index, inplace=True)
df.drop(df[df.director == ""].index, inplace=True)
df.drop(df[df.mainCharactor == ""].index, inplace=True)
df.drop(df[df.introduction == ""].index, inplace=True)

df.to_csv('data/movie_data_predict_noscreen.csv', index=False)

import pandas as pd
df = pd.read_csv('data/movie_data_predict_noscreen.csv')

movie_data_name = pd.read_csv('data/movie_data_Name.csv')

movie_data_id = pd.read_csv('data/movie_data_ID.csv')

movie_data_name.head(1)

movie_data_id.head(1)

# charactor id 중 3550 ~ 3620(70)까지는 없다.
for i in range(5000):
    if(len(movie_data_id[movie_data_id.charactor == i].index) > 0):
        print(i, end=" ")

# 감독 id중 3150 ~ 3220(70)까지는 없다.
for i in range(5000):
    if(len(movie_data_id[movie_data_id.director == i].index) > 0):
        print(i, end=" ")

# 기존에 같은 연출자는 없다.
for char in df.mainCharactor:
    if(len(movie_data_name[movie_data_name.charactor == char].index) > 0):
        print(char, end="  ")

# 기존에 같은 감독자는 없다.
for dir in df.director:
    if(len(movie_data_name[movie_data_name.charactor == dir].index) > 0):
        print(dir, end="  ")

# 같은 애니매이션이 감독과 연출이 같다
df[df.mainCharactor.duplicated() == True]

# charactor id 중 3550 ~ 3620(70)까지는 없다.
df['mainCharactorID'] = 0
charactor = {}
i = 3550
for char in df.mainCharactor:
    try:
        charactor[char]
    except:
        charactor[char] = i
    i += 1
for idx in df.index:
    df.mainCharactorID[idx] = charactor[df.mainCharactor[idx]]

# 감독 id중 3150 ~ 3220(70)까지는 없다.
df['directorID'] = 0
direc = {}
i = 3150
for dir in df.director:
    try:
        direc[dir]
    except:
        direc[dir] = i
    i += 1
for idx in df.index:
    df.directorID[idx] = direc[df.director[idx]]

- 공연실황 : 11, 22/ 뮤지컬 : 30, 38 -> 지우기 || 멜로/로맨스 : 1, 7, 45, 48 -> 로맨스

df.drop(df[df.genres.apply(lambda x: '공연실황' in x)].index, inplace=True)
df.drop(df[df.genres.apply(lambda x: '뮤지컬' in x)].index, inplace=True)

df[df.genres.apply(lambda x: '멜로/로맨스' in x)].genres

import ast
df.genres = df.genres.apply(lambda x: ast.literal_eval(x))

genres_list = ["액션", "모험" ,"판타지", "가족", "애니메이션", "코미디", "SF", "드라마", "공포", "스릴러", "범죄", "미스터리", "음악", "역사", "TV 영화", "전쟁", "멜로/로맨스", "다큐멘터리", "서부"]
df_genre_each = pd.DataFrame(columns=genres_list)
for idx in df.index:
    genres = df.genres[idx]
    # 한 줄 0으로 초기화
    df_genre_each.loc[idx] = 0
    for genre in genres:
        df_genre_each.loc[idx][str(genre)] = 1  # 한 줄 값 부여

df_genre_each.rename(columns= {'멜로/로맨스':'로맨스'},inplace=True)

dataframe = pd.concat([df, df_genre_each], axis=1)

dataframe.to_csv('data/movie_data_predict_noscreenML.csv', index=False)
