==> ML/8.MoviepredictML.ipynb <==
<pre>개봉예정영화 흥행 예측한 파일</pre>

import pandas as pd
from tensorflow.keras.models import load_model

df = pd.read_csv('data/movie_data_predict_noscreenML.csv')
X = df

df['charactorId'] = df['mainCharactorID']
df['directorId'] = df['directorID']

df = df.drop(['poster_path', 'genres', 'director', 'mainCharactor', 'introduction', 'mainCharactorID', 'directorID'], axis=1)

df['charactorId'] = df['charactorId'].apply(lambda x: x/3646905)
df['directorId'] = df['directorId'].apply(lambda x: x/3646897)
df = df.drop(['가족','애니메이션','코미디', '음악', '역사','TV 영화','전쟁','다큐멘터리','서부', '범죄'], axis=1)

df.head(1)

import numpy as np
X_data = np.array(df)

model = load_model('model/Dense_Dropout_nobudget.h5')
yhat = model.predict(X_data)
for i in range(10):
    print(1 if yhat[i] > 0.5 else 0, end=" ")
print()

X['predict'] = yhat

X.head(1)

X.drop(['charactorId', 'directorId'], axis=1, inplace=True)

X.to_csv('data/movie_data_predict_answer.csv', index=False)

print("poster_path", X.poster_path.apply(lambda x: len(x)).max() * 4)
print("genres", X.genres.apply(lambda x: len(x)).max() * 4)
print("director", X.director.apply(lambda x: len(x)).max() * 4)
print("mainCharactor", X.mainCharactor.apply(lambda x: len(x)).max() * 4)
print("introduction", X.introduction.apply(lambda x: len(x)).max() * 4)
